rft: 拆分设置-软件更新
