+++ src/MaaWpfGui/Models/ResourceUpdater.cs
@@ -231,7 +231,7 @@ namespace MaaWpfGui.Models
                 OutputDownloadProgress(LocalizationHelper.GetString("GameResourceUpdated"));
 
                 // 现在用的和自动安装服更新包一个逻辑，看看有没有必要分开
-                if (Instances.SettingsViewModel.AutoInstallUpdatePackage)
+                if (Instances.SettingsViewModel.VersionUpdateDataContext.AutoInstallUpdatePackage)
                 {
                     await Bootstrapper.RestartAfterIdleAsync();
                 }
