+++ src/MaaWpfGui/Views/UserControl/VersionUpdateSettingsUserControl.xaml.cs
@@ -20,6 +20,7 @@ using HandyControl.Controls;
 using HandyControl.Data;
 using MaaWpfGui.Helper;
 using MaaWpfGui.ViewModels.UI;
+using MaaWpfGui.ViewModels.UserControl;
 using Serilog;
 
 namespace MaaWpfGui.Views.UserControl
@@ -51,23 +52,23 @@ namespace MaaWpfGui.Views.UserControl
 
         private void MaaVersionClick(object sender, MouseButtonEventArgs e)
         {
-            CopyToClipboardAsync($"UI Version: {SettingsViewModel.UiVersion}\nCore Version: {SettingsViewModel.CoreVersion}\nBuild Time: {SettingsViewModel.BuildDateTimeCurrentCultureString}");
+            CopyToClipboardAsync($"UI Version: {VersionUpdateSettingsUserControlModel.UiVersion}\nCore Version: {VersionUpdateSettingsUserControlModel.CoreVersion}\nBuild Time: {VersionUpdateSettingsUserControlModel.BuildDateTimeCurrentCultureString}");
         }
 
         private void CoreVersionClick(object sender, MouseButtonEventArgs e)
         {
-            CopyToClipboardAsync("Core Version: " + SettingsViewModel.CoreVersion);
+            CopyToClipboardAsync("Core Version: " + VersionUpdateSettingsUserControlModel.CoreVersion);
             EasterEggs();
         }
 
         private void UiVersionClick(object sender, MouseButtonEventArgs e)
         {
-            CopyToClipboardAsync("UI Version: " + SettingsViewModel.UiVersion);
+            CopyToClipboardAsync("UI Version: " + VersionUpdateSettingsUserControlModel.UiVersion);
         }
 
         private void ResourceVersionClick(object sender, MouseButtonEventArgs e)
         {
-            CopyToClipboardAsync($"Resource Version: {Instances.SettingsViewModel.ResourceVersion}\nResource Time: {Instances.SettingsViewModel.ResourceDateTimeCurrentCultureString}");
+            CopyToClipboardAsync($"Resource Version: {Instances.SettingsViewModel.VersionUpdateDataContext.ResourceVersion}\nResource Time: {Instances.SettingsViewModel.VersionUpdateDataContext.ResourceDateTimeCurrentCultureString}");
         }
 
         private static void CopyToClipboardAsync(string text)
